ml/jk R-189 15127 - #23 Enhanced Validators.countactiveDependencies to return long instead of int, adjusted relevant test.
git-svn-id: http://kid/svn/trident-genesis/trunk@5014 50ee8237-08a1-5d4c-b143-727193748607

+++ platform-dao/src/test/java/ua/com/fielden/platform/validators/CountActiveDependenciesTest.java
@@ -28,7 +28,7 @@ public class CountActiveDependenciesTest extends AbstractDomainDrivenTestCase {
         final TgCategory cat1 = new_(TgCategory.class, "Cat1");
         assertNotNull(cat1);
 
-        final int count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat1, coAggregates);
+        final long count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat1, coAggregates);
         assertEquals(0, count);
     }
 
@@ -39,7 +39,7 @@ public class CountActiveDependenciesTest extends AbstractDomainDrivenTestCase {
         final TgCategory cat1 = ao(TgCategory.class).findByKey("Cat1");
         assertNotNull(cat1);
 
-        final int count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat1, coAggregates);
+        final long count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat1, coAggregates);
         assertEquals(2, count);
     }
 
@@ -49,7 +49,7 @@ public class CountActiveDependenciesTest extends AbstractDomainDrivenTestCase {
         final TgCategory cat2 = ao(TgCategory.class).findByKey("Cat2");
         assertNotNull(cat2);
 
-        final int count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat2, coAggregates);
+        final long count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat2, coAggregates);
         assertEquals(1, count);
     }
 
@@ -59,7 +59,7 @@ public class CountActiveDependenciesTest extends AbstractDomainDrivenTestCase {
         final TgCategory cat3 = ao(TgCategory.class).findByKey("Cat3");
         assertNotNull(cat3);
 
-        final int count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat3, coAggregates);
+        final long count = Validators.countActiveDependencies(domainProvider.entityTypes(), cat3, coAggregates);
         assertEquals(0, count);
     }
